feat: exclude processing jobs sidebar
